git-svn-id: http://192.168.20.11/sequoiadb/trunk@17299 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsCommand.cpp
@@ -612,7 +612,7 @@ namespace engine
                rc = SDB_INVALIDARG ;
                goto error ;
             }
-            _level = ( CLS_REELECTION_LEVEL )e.Number() ;
+            _level = ( CLS_REELECTION_LEVEL )((INT32)e.Number()) ;
          }
       }
       catch ( std::exception &e )
